working page render of login, signup, and homepage
