+++ public/javascript/login.js
@@ -1,12 +1,13 @@
-const loginForm = async (event) => {
+const loginFormHandler = async (event) => {
     //stop the browser from submitting the form immediately so we can add in the api response
     event.preventDefault();
 
     //collect values from the login form
     const email = document.querySelector("#email-login").value.trim();
     const password = document.querySelector("#password-login").value.trim();
-    // const errorMessage = document.querySelector(".error");
+    
     console.log(email, password);
+    
     if (email && password) {
         //send the email and password to the server with a post request
         const response = await fetch ("/api/users/login", {
@@ -21,42 +22,15 @@ const loginForm = async (event) => {
         } else {
             alert("Your login was unsuccessful, please try again");
         }
-        console.log("message");
+        console.log(response);
     }
 };
 
 
 
-// sign up with us
-const signupForm = async(event) => {
-    event.preventDefault();
-    
-    //collect data from the form
-    //no id's here yet for first or last name, do we need them?
-    const lastName = document.querySelector("#lastName").value.trim();
-    const firstName = document.querySelector("first-name").value.trim();
-    const email = document.querySelector("#email-signup").value.trim();
-    const password = document.querySelector("#password").value.trim();
-    
-    if (email && password) {
-        //send data to the server
-        const response = await fetch ("/api/users", {
-            method: "POST",
-            body: JSON.stringify({lastName, firstName, email, password}),
-            headers: { "Content-Type": "application/json" },
-        });
-        
-        if (response.ok) {
-            document.location.replace("/api/users");       
-        } else {
-            alert("Failed to sign up");
-        }
-        console.log(response.statusText);
-    }
-    
-};
 
-document.querySelector(".login-form").addEventListener("submit", loginForm);
 
-document.querySelector(".signup-form").addEventListener("submit", signupForm);
+document.querySelector(".login-form").addEventListener("submit", loginFormHandler);
+
+
