+++ server.js
@@ -3,7 +3,7 @@ const express = require("express");
 const session = require("express-session");
 const exphbs = require("express-handlebars");
 const routes = require("./controllers");
-// const helpers = require("./utils/helpers");//we arent using any helper right now
+const helpers = require("./utils/helpers"); 
 
 const sequelize = require("./config/connection");
 const SequelizeStore = require("connect-session-sequelize")(session.Store);
@@ -11,7 +11,8 @@ const SequelizeStore = require("connect-session-sequelize")(session.Store);
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-const hbs = exphbs.create();
+//use the helper files in utils
+const hbs = exphbs.create(); //helper object in the callback
 
 const oneDay = 100 * 60 * 60 * 24;
 const sess = {
