+++ views/signup.handlebars
@@ -37,3 +37,4 @@
   </div>
 </div>
 
+<script src="./javascript/signup.js"></script>
